requirements.txt
Flask==2.3.3
pandas==2.2.2
psycopg2-binary==2.9.9
openpyxl==3.1.2
pdfkit==1.0.0
Jinja2==3.1.3
gunicorn==21.2.0
weasyprint
